Removed redundant casts.

+++ src/main/java/org/newdawn/slick/Graphics.java
@@ -1709,7 +1709,7 @@ public class Graphics {
             buffer = BufferUtils.createFloatBuffer(18);
             stack.add(buffer);
         } else {
-            buffer = (FloatBuffer) stack.get(stackIndex);
+            buffer = stack.get(stackIndex);
         }
 
         GL.glGetFloat(SGL.GL_MODELVIEW_MATRIX, buffer);
@@ -1732,7 +1732,7 @@ public class Graphics {
         predraw();
 
         stackIndex--;
-        FloatBuffer oldBuffer = (FloatBuffer) stack.get(stackIndex);
+        FloatBuffer oldBuffer = stack.get(stackIndex);
         GL.glLoadMatrix(oldBuffer);
         sx = oldBuffer.get(16);
         sy = oldBuffer.get(17);
